requirements.txt
langchain-google-genai==2.0.9
python-dotenv==1.0.1
streamlit==1.41.1